requirements.txt
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
argcomplete==3.6.2
boto3==1.38.36
botocore==1.38.36
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
distro==1.9.0
eval-type-backport==0.2.2
fasta2a==0.2.18
fastavro==1.11.1
filelock==3.18.0
fsspec==2025.5.1
google-auth==2.40.3
google-genai==1.20.0
griffe==1.7.3
groq==0.28.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib-metadata==8.7.0
jiter==0.10.0
jmespath==1.0.1
logfire-api==3.19.0
markdown-it-py==3.0.0
mcp==1.9.4
mdurl==0.1.2
mistralai==1.8.2
openai==1.86.0
opentelemetry-api==1.34.1
packaging==25.0
prompt-toolkit==3.0.51
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.6
pydantic-ai==0.2.18
pydantic-ai-slim==0.2.18
pydantic-core==2.33.2
pydantic-evals==0.2.18
pydantic-graph==0.2.18
pydantic-settings==2.9.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.4
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.47.0
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
websockets==15.0.1
zipp==3.23.0
